feat(wasm4games): expand to 70 patterns (Phase 2) + pin new GOLDEN_CORPUS_DIGEST
Adds 50 new branchless game-pattern kernels (ids 21-70) across 10 families:
Pathfinding, Procedural Generation, Economy/Progression, Narrative/Dialogue,
Camera, Audio, Multiplayer/Network, DfLSS/Quality, Engine Bridge, AI Agent/Benchmark.

Each kernel follows the ggen-generated convention: GENERATED banner, branchless
fn(u64,u64)->u64 body over bcinr-logic primitives, _reference oracle, 3 mutants,
proptest equivalence/cf1/cf2/cf3/boundaries tests, and bench module.

Integration changes:
- src/patterns/mod.rs: wired all 70 pub mod + pub use + 10 new ObjectKind consts
  (NAV_NODE–EPISODE, 0x0016–0x001F) + 50 new OBJ_ slices + 50 PATTERN_REGISTRY entries
  (event codes 0x0053–0x0084)
- src/corpus.rs: extended dispatch() arms 21–70; re-pinned GOLDEN_CORPUS_DIGEST to
  0x2D26_7F72_6A8B_F791 (70 patterns × 4 probes); updated registry count to 70

Bug fixes discovered during integration:
- action_mask_applied: mask idx to 5 bits (shift overflow on probe u64::MAX)
- biome_class_selected / condition_flag_evaluated: rank_u64 is inclusive-pos;
  switched kernel to mask-below approach to match reference (strictly-below count)
- fov_adjusted: sign-magnitude decode was wrong (bit15 gives 0/1, not 0/-1);
  fixed branchless formula to (1 - 2*bit15)*mag; guarded cf tests with min<=max
- camera_distance_clamped / terrain_height_quantized: std::clamp panics when min>max;
  added prop_assume!(min<=max) guards to cf tests
- All clippy -D warnings clean: unnecessary_cast, identity_op, doc_overindented_list_items,
  manual_clamp across 14 files; stray .rlib files from repo root removed

Law-state: wasm4games = VERIFIED_UNDER_SCOPE (70 patterns); corpus pinned and reproduced.
bcinr workspace crown = BLOCKED_BY_PRE_EXISTING_CONTRACT_GATE_RESIDUAL (unchanged, 17 files).

+++ crates/wasm4games/src/patterns/camera_shake_applied.rs
@@ -24,10 +24,10 @@ use bcinr_logic::patterns::integrity_receipt::DeterministicSubstrateReceipt;
 #[inline]
 #[must_use]
 pub fn camera_shake_applied(state: u64, input: u64) -> u64 {
-    let tick = (state & 0xFFFF) as u64;
-    let seed = ((state >> 16) & 0xFFFF) as u64;
-    let max_amp = (input & 0xFF) as u64;
-    let intensity = ((input >> 8) & 0xFF) as u64;
+    let tick = state & 0xFFFF;
+    let seed = (state >> 16) & 0xFFFF;
+    let max_amp = input & 0xFF;
+    let intensity = (input >> 8) & 0xFF;
     let mut r = DeterministicSubstrateReceipt::new();
     r.record(tick, seed, intensity);
     let hash = r.finalize();
@@ -47,10 +47,10 @@ mod tests {
     use proptest::prelude::*;
 
     fn camera_shake_applied_reference(state: u64, input: u64) -> u64 {
-        let tick = (state & 0xFFFF) as u64;
-        let seed = ((state >> 16) & 0xFFFF) as u64;
-        let max_amp = (input & 0xFF) as u64;
-        let intensity = ((input >> 8) & 0xFF) as u64;
+        let tick = state & 0xFFFF;
+        let seed = (state >> 16) & 0xFFFF;
+        let max_amp = input & 0xFF;
+        let intensity = (input >> 8) & 0xFF;
         let mut r = DeterministicSubstrateReceipt::new();
         r.record(tick, seed, intensity);
         let hash = r.finalize();
@@ -102,7 +102,10 @@ mod tests {
         let s1 = 0u64; // tick=0, seed=0
         let s2 = 1u64; // tick=1, seed=0
         let inp = 10u64 | (128u64 << 8); // max_amp=10, intensity=128
-        assert_eq!(camera_shake_applied(s1, inp), camera_shake_applied_reference(s1, inp));
+        assert_eq!(
+            camera_shake_applied(s1, inp),
+            camera_shake_applied_reference(s1, inp)
+        );
         assert_ne!(camera_shake_applied(s1, inp), camera_shake_applied(s2, inp));
         // Idempotency: same inputs twice -> same output.
         assert_eq!(camera_shake_applied(s2, inp), camera_shake_applied(s2, inp));
